+++ administrator/components/com_loginguard/src/Controller/DisplayController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace LoginGuard\Component\LoginGuard\Administrator\Controller;
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\MVC\Controller\BaseController;
+
+final class DisplayController extends BaseController
+{
+    protected $default_view = 'attempts';
+}